storage types optimized

+++ contracts/interfaces/IComparing.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IComparing {
-    function compare(uint value1, uint value2) external view returns (bool);
+    function compare(uint64 value1, uint64 value2) external view returns (bool);
 }
